% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{ChrawExperimentFromSampleFile}
\alias{ChrawExperimentFromSampleFile}
\title{Constructor of ChrawExperiments from a sample sheet}
\usage{
ChrawExperimentFromSampleFile(sampleFile, genome, minimalMetadata = TRUE)
}
\arguments{
\item{sampleFile}{Path to the sample sheet file.}

\item{genome}{Reference genome can be: mm9, mm10, hg19, hg38, rn6.}

\item{minimalMetadata}{Logical, indicating whether all metadata from the
sample sheet file should be kept or only a set of essential columns.}
}
\value{
A ChrawExperiment object.
}
\description{
This function builds a ChrawExperiment object from a
pre-define
sample sheet containing following information:
\enumerate{
\item sampleName - Sample names.
\item bamFile - Path to alignment file in .bam format.
\item bigwigFile - Path to signal tracks in .bigwig format.
\item peakFile - Path to peaks in either bed, narrowPeaks or broadPeak format.
5-) Any additional sample information.
The ChrawExperiment object is an extension of the MultiAssayExperiment
object specifically designed for the analysis of chromatin experiments.
The ChrawExperiment object aims at being the central data structure for
downstream analyses.
}
}
\examples{
sampleFile <- system.file("files", "sampleSheet.txt", package="chraw",
                         mustWork=TRUE)
sampleDF <- as.data.frame(read.csv(sampleFile, sep = "\t", header=TRUE))

inst_dir <- system.file("files/", package="chraw")
sampleDF$bamFile <- gsub("inst/files", inst_dir, sampleDF$bamFile)
sampleDF$bigwigFile <- gsub("inst/files", inst_dir, sampleDF$bigwigFile)
sampleDF$peakFile <- gsub("inst/files", inst_dir, sampleDF$peakFile)
sampleDF$qcFile <- gsub("inst/files", inst_dir, sampleDF$qcFile)

tmpSheet <- tempfile(fileext = ".txt")
write.table(sampleDF, tmpSheet, sep = "\t", quote = FALSE, row.names = FALSE)

ce <- ChrawExperimentFromSampleFile( tmpSheet, genome = 'rn6')

}
