% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_data.R
\docType{data}
\name{ce_chipseq}
\alias{ce_chipseq}
\alias{ce_atac}
\alias{ce_rat}
\alias{ce_examples}
\alias{chromHMMDict}
\title{ChrawExperiment example objects}
\format{
ChrawExperiment objects

An object of class \code{ChrawExperiment} of length 1.

An object of class \code{ChrawExperiment} of length 2.

An object of class \code{ChrawExperiment} of length 3.

An object of class \code{data.frame} with 28 rows and 2 columns.
}
\usage{
ce_chipseq

ce_atac

ce_rat

ce_examples

chromHMMDict
}
\value{
\code{ce_atac}, \code{ce_chipseq}, \code{ce_rat} and \code{ce_examples} are
ChrawExperiment objects; \code{chromHMMDict} is a data frame mapping the
ChromHMM state names to simplified labels.
}
\description{
ChrawExperiment example objects used throughout the vignette.
}
\examples{
data(ce_atac)
ce_atac

data(chromHMMDict)
head(chromHMMDict)

}
\keyword{data}
\keyword{datasets}
