% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting_genome.R
\name{importAndAverage}
\alias{importAndAverage}
\title{Constructor of coverage data}
\usage{
importAndAverage(
  object,
  includeSamples,
  plottingRegion,
  resolution = 100,
  smoothIter = 2
)
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{includeSamples}{Vector of strings with sample names to be included
in coverage calculation.}

\item{plottingRegion}{A GRanges object with the region to be visualized.}

\item{resolution}{Numerical value for the resolution of the plot. Default
value is 100.}

\item{smoothIter}{Number of iterations to smooth the plot. Default value
is 2.}
}
\value{
Dataframe of coverage data
}
\description{
This function creates a list of \code{GRanges} objects with
the average coverage from
samples of a ChrawExperiment object. \strong{NOTE}: samples names should end
with replicate information
separated by a "_" to create the correct replicate groups.
}
\examples{
data(ce_examples)
ce_examples <- rewrite_paths(ce_examples)

includeSamples <- rownames(colData(ce_examples))[seq_len(3)]
plottingRegion <- GRanges("chr6:67090000-67170000")
covData <- importAndAverage(ce_examples,includeSamples,plottingRegion )

}
