% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiomics.R
\name{plotDiffScatter}
\alias{plotDiffScatter}
\title{Scatterplot comparison of log2(FC) from differential analysis.}
\usage{
plotDiffScatter(
  object,
  experimentNames,
  contrasts,
  filterAnnoCols = NULL,
  filterValues = NULL,
  assignFunc = findOverlaps,
  geneAnnoCols = "nearest_TSS",
  colorByFDR = "mean"
)
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{experimentNames}{Vector with two count experiment names, normally
added using the \code{addCountExperiment} function.}

\item{contrasts}{Vector with two contrast names, used to compute diff peak
statistics using the \code{testForDiffSignal} function.}

\item{filterAnnoCols}{Vector with one/two column names in \code{rowData} for
peak filtering, e.g. simple_annotation added by the
\code{annotateExperimentRegions} function.}

\item{filterValues}{Vector with one/two values to bes used to filter the
peaks, e.g. in case of simple_annotation it can be 'Promoter', 'Intron'
etc.}

\item{assignFunc}{GenomicRanges function to compare peak sets. Add the
moment supported functions include \code{findOverlaps}, \code{precede}, \code{follow} and
\code{nearest}.}

\item{geneAnnoCols}{If one of the experiments is an RNA-seq dataset, the
param \code{geneAnnoCols} must be a character vector that matches a column name
of \code{mcols()} of the other experiment.}

\item{colorByFDR}{A string indicating how to color the points. Values can
be either 'mean' (default) to color the point according to the mean
q-value (FDR) across experiments, \code{exp1} and \code{exp2} to color by the
q-value of individual experiments or 'none'.}
}
\value{
Scatterplot comparing the log2(FC) of two diff. contrasts/region
sets and/or experiments.
}
\description{
Assign peaks to peaks either within the same experiment or
across experiments using an assignment function such as \code{findOverlaps},
\code{nearest}, etc.
}
\examples{
data(ce_examples)
ce_examples <- rewrite_paths(ce_examples)

# Perform diff testing for the scatterplot
contrasts <- list(contrast01 = c("condition", "agonist_3h", "CTRL_3h"),
                 contrast02 = c("condition", "agonist_27h", "CTRL_27h"))

ce_examples <- testForDiffSignal(object = ce_examples,
                           experimentName = "Peaks",
                           design = ~condition,
                           contrasts = contrasts)

# Compare diff signal in promoters and nearby enhancers
plot1 <- plotDiffScatter(object = ce_examples,
               contrasts =  c('contrast01', 'contrast01'),
               experimentNames = c('Peaks', 'Peaks'),
               filterAnnoCols = c('simple_annotation', 'simple_annotation'),
               filterValues = c('Promoter', 'Distal Intergenic'),
               assignFunc = GenomicRanges::nearest)

# Compare diff signal in all peaks for two different contrasts
plot2 <- plotDiffScatter(object = ce_examples,
               contrasts =  c('contrast01', 'contrast02'),
               experimentNames = c('Peaks', 'Peaks'))

}
