% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/utils_chraw.R
\docType{methods}
\name{pipeline}
\alias{pipeline}
\alias{referenceGenome}
\alias{referenceGenome,ChrawExperiment-method}
\alias{pipeline,ChrawExperiment-method}
\title{Accessors for the metadata of a ChrawExperiment object}
\usage{
pipeline(object, ...)

\S4method{referenceGenome}{ChrawExperiment}(x)

\S4method{pipeline}{ChrawExperiment}(object)
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{...}{Parameters passed to object-specific methods.}

\item{x}{A ChrawExperiment object.}
}
\value{
A character vector of length 1 with the reference genome assembly
(\code{referenceGenome()}) or the pipeline name (\code{pipeline()}).
}
\description{
\code{referenceGenome()} returns the reference genome assembly that
the experiments of a \linkS4class{ChrawExperiment} object were aligned to, and
\code{pipeline()} returns the name of the pipeline that produced them. Both are
recorded when the object is built and are used across the package to select
the matching annotation resources.
}
\examples{
data(ce_examples)

referenceGenome( ce_examples )
pipeline( ce_examples )

}
