% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{rewrite_paths}
\alias{rewrite_paths}
\title{Function to rewrite paths in example datasets}
\usage{
rewrite_paths(ce)
}
\arguments{
\item{ce}{A ChrawExperiment object}
}
\value{
The same ChrawExperiment object with the \code{bamFile},
\code{bigwigFile}, \code{peakFile} and \code{qcFile} columns rewritten to the
installed location of the example files.
}
\description{
This function rewrites the paths from the bam, bigwig, peak and qc files
available as test datasests for the package, allowing for the package to
find all the necessary data to run test code.
}
\examples{
data(ce_atac)
ce_atac <- rewrite_paths(ce_atac)
head( colData(ce_atac)$bamFile )

}
