% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tidy.R
\name{tidyChrawExperiment}
\alias{tidyChrawExperiment}
\title{Function to tidy a given experiment from a Chraw MultiAssayExperiment.}
\usage{
tidyChrawExperiment(object, experimentName, assayName = "counts")
}
\arguments{
\item{object}{A ChrawExperiment object}

\item{experimentName}{The name of the count experiment, normally added
using the \code{addCountExperiment} function.}

\item{assayName}{Name of the assay (default: counts) which should be
converted from wide to long format.}
}
\value{
Tidy dataframe of experiment in ChrawExperiment object
}
\description{
Function to tidy a given experiment from a Chraw MultiAssayExperiment.
}
\examples{
data('ce_examples')

ce_examples <- rewrite_paths(ce_examples)

tidyChrawExperiment(ce_examples, experimentName = 'Peaks')
tidyChrawExperiment(ce_examples, experimentName = 'Peaks', assayName = 'cpms')

}
