% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiAssayExperiment.R,
%   R/utils_chraw.R
\name{addCountExperiment}
\alias{addCountExperiment}
\alias{c,ChrawExperiment-method}
\title{Adds a count experiment to a ChrawExperiment object}
\usage{
addCountExperiment(
  object,
  regions,
  name,
  includeSamples = rownames(colData(object)),
  dedup = FALSE,
  isPairedEnd = TRUE,
  normalizeCPM = TRUE,
  BPPARAM = SerialParam(),
  ...
)

\S4method{c}{ChrawExperiment}(x, ..., sampleMap = NULL, mapFrom = NULL)
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{regions}{A named GenomicRegions with the genomic coordinates of the regions to count.}

\item{name}{A string specifying the name of the new experiment.}

\item{includeSamples}{A character vector specifying the samples to be included in the new experiment. These names should match with \code{rownames(colData(object))}.}

\item{dedup}{Logical indicating whether to use the bam files without PCR duplicates or not.}

\item{isPairedEnd}{Logical indicating whether reads are paired-end or not.}

\item{normalizeCPM}{Logical indicating whether counts should be normalized to CPMs or not.}

\item{BPPARAM}{A BiocParallel instance.}

\item{...}{additional arguments passed to \code{featureCounts()}}

\item{x}{A \code{ChrawExperiment} object.}

\item{sampleMap}{\code{c} method: a \code{sampleMap} \code{list} or
\code{DataFrame} to guide merge}

\item{mapFrom}{Either a \code{logical}, \code{character}, or \code{integer}
vector indicating the experiment(s) that have an identical colname order as
the experiment input(s). If using a character input, the name must match
exactly.}
}
\value{
A ChrawExperiment object with a new experiment added.
}
\description{
This function inputs a ChrawExperiment and counts the number
of read fragments overlapping with a set of specified genomic ranges. The read
counting can be done either for all samples or only a subset of them. The results
are formatted as a SummarizedExperiment object and saved as part of the
ChrawExperiment object.
}
