% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiAssayExperiment.R
\name{addFragmentLengthDist}
\alias{addFragmentLengthDist}
\title{Add fragment length distribution per sample to ChrawExperiment.}
\usage{
addFragmentLengthDist(
  object,
  maxFragLength = 750,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{maxFragLength}{A numeric value specifying the maximum fragment length to consider.}

\item{BPPARAM}{A BiocParallel instance for parallelization.}

\item{...}{Further options passed to \code{csaw::getPESizes()}}
}
\value{
ChrawExperiment object with "FragLengthDist" experiment
}
\description{
This function calculates the length distribution
of the sequenced fragments and adds these information as an experiment
to the ChrawExperiment object.
}
