% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multiAssayExperiment.R
\name{addRNASeqExperiment}
\alias{addRNASeqExperiment}
\title{Add an RNA-seq experiment to a ChrawExperiment object}
\usage{
addRNASeqExperiment(object, se, name = NULL, identifierType = "ENSEMBL")
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{se}{Either a 'RangedSummarizedExperiment' or a 'DESeqDataSet' object.}

\item{name}{A string specifying the name of the new experiment.}

\item{identifierType}{A character string specifying the identifier type. Can be 'ENSEMBL', 'ENTREZID' or 'SYMBOL'.}
}
\value{
A ChrawExperiment object with an RNA-seq experiment added.
}
\description{
This function inputs a ChrawExperiment object
and either a RangedSummarizedExperiment or a DESeqDataSet object
containing RNA-seq count data. The function integrates the
RNA-seq count data into the ChrawExperiment container. It also
maps ENTREZ, ENSEMBL and SYMBOL identifiers to enable
integration with the chromatin datasets.
}
