% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chraw-package.R
\docType{package}
\name{chraw-package}
\alias{chraw}
\alias{chraw-package}
\title{chraw: analysis of chromatin and multi-omics datasets}
\description{
\code{chraw} analyses chromatin and multi-omic experiments. It extends the
\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment} class into a
\linkS4class{ChrawExperiment} container that additionally records the reference
genome and the processing pipeline of the experiment, and it collapses each
major analysis step (quality control, read counting, differential testing,
peak annotation, enrichment and visualisation) into a single function call.

The typical workflow is:
\enumerate{
\item Build the container with \code{\link[=ChrawExperimentFromSampleFile]{ChrawExperimentFromSampleFile()}} or
\code{\link[=ChrawExperimentFromDataFrame]{ChrawExperimentFromDataFrame()}}.
\item Inspect the quality of the experiments with \code{\link[=plotMappingStats]{plotMappingStats()}},
\code{\link[=plotPCRDuplicateStats]{plotPCRDuplicateStats()}}, \code{\link[=plotChrMStats]{plotChrMStats()}}, \code{\link[=plotFripStats]{plotFripStats()}},
\code{\link[=plotFracReadsInAnnot]{plotFracReadsInAnnot()}} and \code{\link[=plotFragmentLengthDist]{plotFragmentLengthDist()}}.
\item Define the regions of interest with \code{\link[=importNarrowPeaks]{importNarrowPeaks()}} or
\code{\link[=getGenomicBins]{getGenomicBins()}} and count reads over them with
\code{\link[=addCountExperiment]{addCountExperiment()}}.
\item Test for differential signal with \code{\link[=testForDiffSignal]{testForDiffSignal()}} and retrieve the
results with \code{\link[=pullDiffResults]{pullDiffResults()}}.
\item Annotate the regions with \code{\link[=annotateExperimentRegions]{annotateExperimentRegions()}} and test
annotation enrichments with \code{\link[=enrichAnno]{enrichAnno()}}.
}

See \code{vignette("chraw", package = "chraw")} for a worked example on both
ATAC-seq and ChIP-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Novartis/chraw}
  \item Report bugs at \url{https://github.com/Novartis/chraw/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Rosa Junior \email{ricardo.rosa@hc.fm.usp.br}

Authors:
\itemize{
  \item Alejandro Reyes \email{alejandro.reyes_quiroz@novartis.com}
  \item Sebastian Steinhauser \email{sebastian.steinhauser@novartis.com}
  \item Christian Kolter \email{christian.kolter@novartis.com}
}

Other contributors:
\itemize{
  \item Novartis Pharma AG [copyright holder]
}

}
\keyword{internal}
