% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{ChrawExperiment-class}
\alias{ChrawExperiment-class}
\title{ChrawExperiment class}
\description{
This class is an extension of the MultiAssayExperiment class,
containing also information on the reference genome from the experiment
and the pipeline used to create the experiments stored in the objects.
}
