% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chraw.R
\name{computeFragLengthDist}
\alias{computeFragLengthDist}
\title{Compute fragment length distribution from a paired-end *.bam file.}
\usage{
computeFragLengthDist(bam, ...)
}
\arguments{
\item{bam}{A character string specifying the file paths to bam file.}

\item{...}{Additional parameters passed to \code{getPESizes()}.}
}
\value{
Data frame with fragment lenghts and thei frequency
}
\description{
Compute fragment length distribution from a paired-end *.bam file.
}
