% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_data.R
\docType{data}
\name{ce_chipseq}
\alias{ce_chipseq}
\alias{ce_atac}
\alias{ce_rat}
\alias{ce_examples}
\alias{chromHMMDict}
\title{ChrawExperiment example objects}
\format{
ChrawExperiment objects

An object of class \code{ChrawExperiment} of length 1.

An object of class \code{ChrawExperiment} of length 2.

An object of class \code{ChrawExperiment} of length 3.

An object of class \code{data.frame} with 28 rows and 2 columns.
}
\usage{
ce_chipseq

ce_atac

ce_rat

ce_examples

chromHMMDict
}
\description{
ChrawExperiment example objects used throughout the vignette.
}
\keyword{data}
\keyword{datasets}
