% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichment.R, R/functions_plotting.R
\name{enrichAnno}
\alias{enrichAnno}
\alias{plotEnrichResults}
\title{Function to test enrichment of peak annotation in up/down regulated peaks.}
\usage{
enrichAnno(
  object,
  experimentName,
  contrastName,
  annoName,
  padjThres = 0.05,
  log2FcThrs = 0,
  foreground = "up",
  background = "all",
  ...
)

plotEnrichResults(enrichRes, pvalueThres = 0.05)
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{experimentName}{The name of the count experiment to test, normally added using the \code{addCountExperiment} function.}

\item{contrastName}{The name of the differential peak contrast used in the \code{testForDiffSignal} function.}

\item{annoName}{The name of the peak annotation added with \code{annotateExperimentRegions} function. This name will be searched in the colnames of the \code{rowData()} of the 'experimentName'.}

\item{padjThres}{Threshold for the adjusted pvalue from the diff peak analysis with \code{testForDiffSignal}. Default: 0.05.}

\item{log2FcThrs}{Threshold for the log2FC from the diff peak analysis. Default: 0.}

\item{foreground}{Define which peak set showed be tested, either "up" or "down" regulated peaks. Default: 'up'.}

\item{background}{Define which peak set should be used as background in the fisher's exact test, either "all" non diff peaks, "up" and "down". Default: 'all'.}

\item{...}{Additional parameters passed to \code{fisher.test()}.}

\item{enrichRes}{A data.frame contating the annotation enrichment from \code{enrichAnno}.}

\item{pvalueThres}{P-value threshold to adjust the color scale.}
}
\value{
A data.frame containing following columns:
\itemize{
\item{'annoName': Annotation name from the peak rowData.}
\item{'n_anno_fg_peaks': Number of foreground peaks overlapping the annotation.}
\item{'n_anno_bg_peaks': Number of background peaks overlapping the annotation.}
\item{'conf1': Lower boundry of the log2 odds confidence interval from \code{fisher.test}.}
\item{'conf2': Upper boundry of the log2 odds confidence interval from \code{fisher.test}.}
\item{'odds': Log2 odds ratio from the test statistic of \code{fisher.test}.}
\item{'pvalue': pvalue from \code{fisher.test}.}
}
}
\description{
The function computes the enrichment of peak annotation in a
subset of differential peaks using a Fisher's exact test. For example,
the enrichment of chromHMM states in differential peaks compared with all peaks.
}
\examples{
data(ce_examples)
ce_examples <- rewrite_paths(ce_examples)

ce_examples <- testForDiffSignal(
   object=ce_examples, experimentName ="Peaks",
   design = ~condition,
   contrasts = list(
   agonist_3h=c("condition", "agonist_3h", "CTRL_3h")))
ce_examples <- annotateExperimentRegions( ce_examples, "Peaks" )
ftResults <- enrichAnno(ce_examples,"Peaks",contrastName="agonist_3h",annoName="simple_annotation" )

}
\seealso{
fisher.test
}
