% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_diff.R
\name{getDiffResultSummary}
\alias{getDiffResultSummary}
\title{Returns all the results of diffential analyses stored in a ChrawExperiment object.}
\usage{
getDiffResultSummary(object)
}
\arguments{
\item{object}{A ChrawExperiment object.}
}
\value{
A \code{data.frame()} summarizing the ChrawExperiment object that include the results of the differential analysis.
}
\description{
This function loops over the \code{experiments()} that were added using the
\code{addCountData()} function to the ChrawExperiment object, to look for results of
differential testing generated using the \code{testForDiffSignal()}. It returns a
\code{data.frame()} that summarizes the results stored in the ChrawExperiment object.
}
