% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/utils_chraw.R
\docType{methods}
\name{getGenomicBins}
\alias{getGenomicBins}
\alias{getGenomicBins,BSgenome-method}
\alias{getGenomicBins,ChrawExperiment-method}
\title{Partition a reference genome in bins}
\usage{
getGenomicBins(object, ...)

\S4method{getGenomicBins}{BSgenome}(object, binSize, onlyStandardChromosomes = TRUE)

\S4method{getGenomicBins}{ChrawExperiment}(object, binSize, onlyStandardChromosomes = TRUE)
}
\arguments{
\item{object}{A BSgenome object or a ChrawExperiment object}

\item{...}{Parameters passed to object-specific methods.}

\item{binSize}{A numeric value specifying the size of the bins desired.}

\item{onlyStandardChromosomes}{Logical indicating whether to keep only standard chromosomes. If 'FALSE' chromosome patches and haplotypes are dropped, as well as the mitochrondrial and 'Y' chromosomes.}
}
\description{
This method inputs either a ChrawExperiment or a BSgenome and returns
a GenomicRanges object with equally sized genomic bins across the genome.
}
