% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_peaks.R
\name{importNarrowPeaks}
\alias{importNarrowPeaks}
\title{Import peaks from a ChrawExperiment object}
\usage{
importNarrowPeaks(
  object,
  includeSamples = rownames(colData(object)),
  peakType = "perSample",
  merge = FALSE
)
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{includeSamples}{A character vector specifying the samples to be included in the new experiment. These names should match with \code{rownames(colData(object))}.}

\item{peakType}{A character string. Valid values are 'perSample' to import the peaks called per sample, "idr" to import the IDR peak calls and "overlap" to}

\item{merge}{A logical flag indicating whether peaks should be merged.}
}
\value{
A ChrawExperiment object with a new experiment added.
}
\description{
This function inputs a ChrawExperiment and imports
the peak for the specified samples. The output is a GRangesList
by default.
}
