% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_peaks.R
\name{importNarrowPeaksFromFile}
\alias{importNarrowPeaksFromFile}
\title{Import peaks from *.narrowPeak file into GRange object}
\usage{
importNarrowPeaksFromFile(narrowPeak.file)
}
\arguments{
\item{narrowPeak.file}{File path to the narrow peaks bed file.}
}
\value{
GRange object
}
\description{
Import peaks from *.narrowPeak file into GRange object
}
