% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chraw.R
\docType{methods}
\name{indexBam}
\alias{indexBam}
\alias{indexBam,ChrawExperiment-method}
\title{Index the bam files from a ChrawExperiment object}
\usage{
\S4method{indexBam}{ChrawExperiment}(files, dedup = FALSE, BPPARAM = SerialParam(), force = FALSE)
}
\arguments{
\item{files}{A ChrawExperiment object.}

\item{dedup}{Logical indicating whether to index the deduplicated bam files or the full bam files.}

\item{BPPARAM}{A BiocParallel instance.}

\item{force}{Logical indicating whether indexes should be re-created in case they exist.}
}
\description{
This function defines the S4 method indexBam for the object ChrawExperiment.
This method locates the bam file of a ChrawExperiment and creates the indexes
for the bam files that are not indexed.
}
