% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/functions_plotting.R
\docType{methods}
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotMA}
\alias{plotMA,ChrawExperiment-method}
\alias{plotVolcano,ChrawExperiment-method}
\title{Plots to explore the sample-to-sample correlations between samples.}
\usage{
plotVolcano(object, ...)

\S4method{plotMA}{ChrawExperiment}(object, experimentName, contrastName = NULL, FDR = 0.1, meanFilter = 0)

\S4method{plotVolcano}{ChrawExperiment}(
  object,
  experimentName,
  contrastName = NULL,
  FDR = 0.1,
  meanFilter = 0
)
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{...}{Parameters passed to object-specific methods.}

\item{experimentName}{The name of the count experiment, normally added using the \code{addCountExperiment} function.}

\item{contrastName}{A name of the contrast name to extract. For a list of the contrast names, consider running \code{getDiffResultSummary()}. If missing, all available results will be returned.}

\item{FDR}{An FDR threshold to assess significance.}

\item{meanFilter}{A numeric value to filter from plotting peaks below these threshold.}
}
\description{
These functions document the S4 method plotPCA for
count experiments of a ChrawExperiment object
}
