% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plotMappingStats}
\alias{plotMappingStats}
\alias{plotPCRDuplicateStats}
\alias{plotChrMStats}
\alias{plotFripStats}
\alias{plotFracReadsInAnnot}
\alias{plotFragmentLengthDist}
\title{Functions to plot QC statistics}
\usage{
plotMappingStats(object)

plotPCRDuplicateStats(object)

plotChrMStats(object)

plotFripStats(object)

plotFracReadsInAnnot(object)

plotFragmentLengthDist(object)
}
\arguments{
\item{object}{A ChrawExperiment object.}
}
\description{
These functions parse the QC statistics from the json QC
file produced by the cromwell pipeline. These functions allow users to
plots the number of mapped and unmapped reads for each sample, number of
PCR duplicates per sample, fraction of reads in known annotation,
fraction of reads mapping to chrM, fraction of reads in peaks and
fragment length distributions.
}
\examples{
data(ce_examples)
ce_examples <- rewrite_paths(ce_examples)

plot_ms <- plotMappingStats( ce_examples )
plot_pcr <- plotPCRDuplicateStats( ce_examples )

}
