% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plotVProfile}
\alias{plotVProfile}
\title{Vplot functionality}
\usage{
plotVProfile(
  object,
  regions,
  windowSize = 600,
  returnMatrix = FALSE,
  sampleLabelColumn = NULL,
  includeSamples = NULL,
  BPPARAM = SerialParam(),
  zlim = 0.95,
  rowScaled = FALSE
)
}
\arguments{
\item{object}{A ChrawExperiment object.}

\item{regions}{GRanges objects containing the regions to plot.}

\item{windowSize}{Window size in bp, used for resizing the regions.}

\item{returnMatrix}{Logical, indicating whether the vplot matrix should be returned instead of the plot.}

\item{sampleLabelColumn}{A character string with the column name of \code{colData(object)}. The values of this column will be used as sample labels.}

\item{includeSamples}{A character vector specifying which samples to include. These must be in \code{rownames(colData(object))}.}

\item{BPPARAM}{A BiocParallel instance.}

\item{zlim}{Numeric value between 0 and 1. Specifies the upper quantile value to be considered as maximum value in the color scale. Larger values in the heatmap are set to this max value (using pmin).}

\item{rowScaled}{Logical flag, indicating whether each rows of the heatmaps should be standarized (z-scores) or not.}
}
\value{
Either a long formatted matrix, or a ggplot object.
}
\description{
This function calculates the 'vplot' profile given a
ChrawExperiment object and a set of genomic ranges.
}
