% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{rewrite_paths}
\alias{rewrite_paths}
\title{Function to rewrite paths in example datasets}
\usage{
rewrite_paths(ce)
}
\arguments{
\item{ce}{A ChrawExperiment object}
}
\description{
This function rewrites the paths from the bam, bigwig, peak and qc files
available as test datasests for the package, allowing for the package to
find all the necessary data to run test code.
}
