% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{selectBSgenome}
\alias{selectBSgenome}
\title{Selects the corresponding BSgenome for a ChrawExperiment}
\usage{
selectBSgenome(x)
}
\arguments{
\item{x}{A ChrawExperiment object.}
}
\description{
This function inputs a ChrawExperiment and returns
the respective BSgenome object that corresponds to the experiment.
}
\examples{
data(ce_examples)
ce_examples <- rewrite_paths(ce_examples)

bsgenome <- selectBSgenome( ce_examples )

}
