% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{selectOrgDb}
\alias{selectOrgDb}
\title{Selects the corresponding org.\strong{.}.db package for a ChrawExperiment object}
\usage{
selectOrgDb(x)
}
\arguments{
\item{x}{A ChrawExperiment object.}
}
\description{
This function inputs a ChrawExperiment and returns
the corresponding org.\strong{.}.db object for the experiment.
}
\examples{
data(ce_examples)
ce_examples <- rewrite_paths(ce_examples)

org <- selectOrgDb( ce_examples )

}
