% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{selectTxDb}
\alias{selectTxDb}
\title{Selects the corresponding TxDb package for a ChrawExperiment object}
\usage{
selectTxDb(x)
}
\arguments{
\item{x}{A ChrawExperiment object.}
}
\description{
This function inputs a ChrawExperiment and returns
the corresponding TxDb object for the experiment.
}
\examples{
data(ce_examples)
ce_examples <- rewrite_paths(ce_examples)

txdb <- selectTxDb( ce_examples )

}
